requirements.txt
numpy
sounddevice
piper-tts==1.2.0
onnxruntime==1.17.1
num2words
evdev
pygame
gtts
requests
RPLCD
requests
gtts